pyproject.toml
[project]
name = "fasthtml-htmx-table"
version = "0.1.0"
description = "Example FastHTML with HTMX table"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.15",
]
